# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LMS «Медицина Долголетия» — API
  version: "1.0.0"
  description: |
    API платформы обучения врачей. Авторизация — через httpOnly cookie `token` (JWT),
    выдаётся эндпоинтами `/api/auth/login` и `/api/auth/register`.
    Роли: `student` (врач), `curator`, `admin`, `super_admin`.
servers:
  - url: /api
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
paths:
...
  /invites:
    get:
      tags: [Invites]
      summary: Список активных приглашений (curator+)
      responses: { "200": { description: OK } }
    post:
      tags: [Invites]
      summary: Пригласить по email (роль назначается при первой регистрации на этот email)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, role]
              properties:
                email: { type: string }
                role: { type: string, enum: [student, curator, admin] }
      responses:
        "200": { description: OK }
        "403": { description: "Недостаточно прав для этой роли" }
  /invites/bulk:
    post:
      tags: [Invites]
      summary: Массовое приглашение врачей (из CSV/Excel, распарсенного на фронтенде)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [emails]
              properties:
                emails:
                  type: array
                  items: { type: string }
                  maxItems: 500
      responses:
        "200":
          description: "Итог загрузки"
          content:
            application/json:
              schema:
                type: object
                properties:
                  created: { type: array, items: { type: string } }
                  skipped:
                    type: array
                    items:
                      type: object
                      properties: { email: { type: string }, reason: { type: string } }
  /invites/{email}:
    delete:
      tags: [Invites]
      summary: Отменить приглашение (admin/super_admin)
      parameters:
        - { name: email, in: path, required: true, schema: { type: string } }
      responses: { "200": { description: OK }, "404": { description: "Не найдено" } }
...
  /messages/{studentId}:
    get:
      tags: [Messages]
      summary: История чата с врачом
      parameters: [{ name: studentId, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: OK } }
  /messages:
    post:
      tags: [Messages]
      summary: Отправить сообщение
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [studentId, text]
              properties: { studentId: { type: string }, text: { type: string } }
      responses: { "200": { description: OK } }
...